apps/dataset-publisher/pyproject.toml
[project]
name = "qafiyah-dataset-publisher"
version = "1.0.0"
description = "Publishes Qafiyah classical Arabic poetry dataset to Hugging Face Hub"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Qafiyah Contributors" }]
keywords = ["arabic", "poetry", "dataset", "huggingface", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "datasets>=2.18.0",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "huggingface-hub>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
qafiyah-publish = "publish:main"

[project.urls]
Homepage = "https://qafiyah.com"
Repository = "https://github.com/alwalxed/qafiyah"
Documentation = "https://github.com/alwalxed/qafiyah/tree/main/apps/dataset-publisher"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
